# src/main/resources/swagger/api.yml
...
openapi: '3.0.1'
info:
  description: |
    Activiseek API
  version: "0.0.1"
  title: Activiseek API
  termsOfService: 'http://localhost:8191/terms/'
  contact:
    name: Michael Pollente
    email: pollentem@gmail.com
servers:
  - url: http://localhost:8191
    description: Development server
  - url: https://localhost:8191
    description: Development server with TLS Profile
...
components:
  schemas:
    Card:
      type: object
      properties:
        id:
          type: string
        business:
          $ref: '#/components/schemas/Business'
    Business:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          description: Name of the business
        url:
          type: string
          description: Yelp URL of the business page
        imageUrl:
          type: string
          description: URL of the photo for this business
        location:
          $ref: '#/components/schemas/BusinessLocation'
        priceRange:
          type: string
          description: Price bracket of the restaurant (1 being budget friendly and 4 being the most expensive)
        userRating:
          type: number
          format: float
          description: Rating for this business
        reviewCount:
          type: integer
          description: Number of reviews for this business
    BusinessLocation:
      type: object
      description: Restaurant location details
      properties:
        latitude:
          type: string
          description: The latitude of the business
        longitude:
          type: string
          description: The longitude of the business
        address:
          type: string
          description: Street address of the business
        city:
          type: string
          description: City of the business
        state:
          type: string
          description: State or province of the business
        zipcode:
          type: string
          description: Zip code of the business
        country:
          type: string
          description: Country code of the business
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          example:
            { "message": "Authorization information is missing or invalid." }
...
  securitySchemes:
    jwt:            # arbitrary name for the security scheme
      type: http
      description: JWT Authentication
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
security:
  - jwt: []        # Apply the security globally to all operations
paths:
  /feed:
    get:
      tags:
        - feed
      summary: Refresh the feed
      operationId: getFeed
      parameters:
        - in: query
          name: term
          required: false
          description: Search term
          example: food
          schema:
            type: string
        - in: query
          name: location
          required: false
          description: Geographic area. Required if latitude and longitude is not provided
          example: Manila
          schema:
            type: string
        - in: query
          name: latitude
          required: false
          description: Latitude of the location. Required if location is not provided
          example: 37.123
          schema:
            type: string
        - in: query
          name: longitude
          required: false
          description: Longitude of the location. Required if longitude is not provided
          example: -122.523
          schema:
            type: string
        - in: query
          name: categories
          required: false
          description: Categories to filter the results
          example: korean,bbq
          schema:
            type: string
        - in: query
          name: price
          required: false
          description: Pricing levels to filter the results
          example: 1,2
          schema:
            type: string
        - in: query
          name: attributes
          required: false
          description: Additional filters
          example: hot_and_new
          schema:
            type: string
      responses:
        '200':
          description: Feed successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
              example:
                [{
                   id: "0001",
                   business: {
                     id: "0001",
                     name: "Mang Inasal",
                     url: "https://www.yelp.com/biz/mang-inasal-makati?adjust_creative=J61ctkwXVl7xX3JLAk_Qsw&utm_campaign=yelp_api_v3&utm_medium=api_v3_business_search&utm_source=J61ctkwXVl7xX3JLAk_Qsw",
                     imageUrl: "https://www.yelp.com/biz/mang-inasal-makati?adjust_creative=J61ctkwXVl7xX3JLAk_Qsw&utm_campaign=yelp_api_v3&utm_medium=api_v3_business_search&utm_source=J61ctkwXVl7xX3JLAk_Qsw",
                     location: {
                       latitude: 14.55283,
                       longitude: 121.013718,
                       address: "Makati Avenue",
                       city: "Makati",
                       state: "NCR",
                       zipCode: "1111",
                       country: "PH"
                     },
                     priceRange: "P",
                     userRating: 4.5,
                     reviewCount: 4
                   }
                 }]
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...